// frontend/src/interfaces/Auth.interfaces.ts
export interface Auth {
  accessToken: string
  id: number
  firstName?: string
  lastName?: string
  birthday?: string
  ci?: string
  email?: string
  password?: string
  permissionLevel?: string
  isActive?: boolean
  cellphone?: any
  lastLogin?: string
  createdAt?: string
  updatedAt?: string
}